auth (#10)
* auth

* Add Ronnie as an admin

* Making some UI changes

* Remove console.log

* More stuff

* [Fixed Stuff]
- Fixed authorization for Ronnie on admin page
- Fixed padding

---------

+++ src/app/api/product/[id]/route.ts
@@ -1,9 +1,13 @@
 import { prisma } from '@/db'
 import { ProductAttr } from '@/shared/product-types'
 import { Body, Params } from '@/shared/routes'
-import { NextResponse } from 'next/server'
+import { NextRequest, NextResponse } from 'next/server'
+import { authGuard } from '../../auth-guard'
 
-export async function DELETE(_: Request, { params }: { params: Params<'DELETE /product/[id]'> }) {
+export async function DELETE(req: NextRequest, { params }: { params: Params<'DELETE /product/[id]'> }) {
+  const res = await authGuard(req)
+  if (res) return res
+
   await prisma.product.delete({
     where: {
       id: params.id,
@@ -12,8 +16,11 @@ export async function DELETE(_: Request, { params }: { params: Params<'DELETE /p
   return new NextResponse()
 }
 
-export async function PUT(request: Request, { params }: { params: Params<'PUT /product/[id]'>; body: ProductAttr }) {
-  const body: Body<'PUT /product/[id]'> = await request.json()
+export async function PUT(req: NextRequest, { params }: { params: Params<'PUT /product/[id]'>; body: ProductAttr }) {
+  const body: Body<'PUT /product/[id]'> = await req.json()
+  const res = await authGuard(req)
+  if (res) return res
+
   await prisma.product.update({
     where: {
       id: params.id,
